refactor: Category 컴포넌트 타입스크립트로 변환 및 기본 썸네일 렌더링 함수 추가
